+++ cmd/grafana-agentctl/main.go
@@ -12,8 +12,6 @@ import (
 	"strings"
 	"syscall"
 
-	"gopkg.in/yaml.v2"
-
 	"github.com/grafana/agent/pkg/config"
 	"github.com/grafana/agent/pkg/logs"
 	"github.com/olekukonko/tablewriter"
@@ -65,7 +63,6 @@ func main() {
 		targetStatsCmd(),
 		samplesCmd(),
 		operatorDetachCmd(),
-		templateDryRunCmd(),
 		testLogs(),
 	)
 
@@ -429,40 +426,6 @@ func filterAgentOwners(refs []meta_v1.OwnerReference) (filtered []meta_v1.OwnerR
 		filtered = append(filtered, ref)
 	}
 	return
-}
-
-func templateDryRunCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "template-parse [directory]",
-		Short: "dry run dynamic configuration",
-		Long:  `This will load the dynamic configuration, load configs, run templates and then output the full config as yaml`,
-		Args:  cobra.ExactArgs(1),
-
-		RunE: func(_ *cobra.Command, args []string) error {
-			cmf, err := config.NewDynamicLoader()
-			if err != nil {
-				return err
-			}
-			c := &config.Config{}
-			err = cmf.LoadConfigByPath(args[0])
-			if err != nil {
-				return err
-			}
-			err = cmf.ProcessConfigs(c)
-			if err != nil {
-				return fmt.Errorf("error processing config templates %s", err)
-			}
-
-			outBytes, err := yaml.Marshal(c)
-			if err != nil {
-				return err
-			}
-			fmt.Println(string(outBytes))
-			return nil
-		},
-	}
-
-	return cmd
 }
 
 func testLogs() *cobra.Command {